Edit setup script

+++ setup.sh
@@ -7,11 +7,13 @@ set -e
 git submodule update --init --recursive
 
 # set up FRUT
+if [ -d modules/FRUT/prefix ]; then
+    echo "CMake already installed. To do a fresh install, delete modules/FRUT/prefix/, and run this script again."
+    exit
+fi
+
 cd modules/FRUT/
 rm -Rf build/
 mkdir build && cd build/
 cmake .. -DCMAKE_INSTALL_PREFIX=../prefix -DJUCE_ROOT=../../JUCE
 cmake --build . --target install
-
-# set up JUCE
-# TODO
